requirements.txt
torch==1.1.0
torchvision==0.2.0
numpy
pillow
tensorboardX
plyfile